get order

+++ internal/pkg/models/customers.gen.go
@@ -15,7 +15,7 @@ type Customer struct {
 	FacebookID  string    `gorm:"column:facebook_id;primaryKey" json:"facebook_id"`
 	LastName    string    `gorm:"column:last_name" json:"last_name"`
 	Image       string    `gorm:"column:image" json:"image"`
-	PhoneNumber int32     `gorm:"column:phone_number" json:"phone_number"`
+	PhoneNumber int64     `gorm:"column:phone_number" json:"phone_number"`
 	FirstName   string    `gorm:"column:first_name" json:"first_name"`
 	ID          string    `gorm:"column:id;not null;default:gen_random_uuid()" json:"id"`
 	CreatedAt   time.Time `gorm:"column:created_at;not null;default:CURRENT_TIMESTAMP" json:"created_at"`
